Add tests for union of models which are similar
In context of Union type the decision which deserializer to use is based
on the actual field type, which in turn derived using pydantic parse_obj.
So if pydantic choose a wrong field type on serialization, a wrong
deserializer is chosen on deserialization.

The solution is to switch on `smart_union` flag for such models

See related issue: https://github.com/dapper91/pydantic-xml/issues/72